Move `setup_logger` back
