+++ utilix/__init__.py
@@ -1,6 +1,5 @@
 __version__ = "0.8.5"
 
-from . import log
 from . import config
 
 # try loading config, if it doesn't work then set uconfig to None
@@ -8,10 +7,10 @@ from . import config
 uconfig = config.Config()
 
 if uconfig.is_configured:
-    logger = log.setup_logger(uconfig.logging_level)
+    logger = config.setup_logger(uconfig.logging_level)
 else:
     uconfig = None  # type: ignore
-    logger = log.setup_logger()
+    logger = config.setup_logger()
 
 from .rundb import DB, xent_collection, xe1t_collection
 from .mongo_files import MongoUploader, MongoDownloader, APIUploader, APIDownloader
