+++ utilix/config.py
@@ -1,6 +1,19 @@
 import os
 import configparser
-from .log import setup_logger
+import logging
+
+
+def setup_logger(logging_level="WARNING"):
+    logger = logging.getLogger("utilix")
+    if logger.hasHandlers():
+        logger.handlers.clear()
+    ch = logging.StreamHandler()
+    ch.setLevel(logging_level)
+    logger.setLevel(logging_level)
+    formatter = logging.Formatter("%(asctime)s - %(name)s - %(levelname)s - %(message)s")
+    ch.setFormatter(formatter)
+    logger.addHandler(ch)
+    return logger
 
 
 class EnvInterpolation(configparser.BasicInterpolation):
